change realisation of footer, fixed warnings from console, clean code
