+++ src/components/footer/Footer.tsx
@@ -1,5 +1,8 @@
+import { FooterLink } from '@/components/footer/footerLink/FooterLink';
+import { ListWrapper } from '@/components/footer/ListWrapper';
+import { footerBlocks } from '@/constants/footer';
 import Image from 'next/image';
-import React from "react";
+import React from 'react';
 import Link from "next/link";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
@@ -9,76 +12,46 @@ import {
 } from "@fortawesome/free-brands-svg-icons";
 import styles from "./footer.module.scss";
 
-interface LinkItem {
+export interface LinkItem {
   label: string;
   href: string;
 }
 
-const FooterLink: React.FC<LinkItem> = ({ label, href }) => (
-  <li>
-    <Link href={href} className={styles.listItem}>
-      {label}
-    </Link>
-  </li>
-);
-
 const Footer: React.FC = () => {
-  const linkItems: LinkItem[] = [
-    { label: "About Us", href: "/" },
-    { label: "Terms & Conditions", href: "/" },
-    { label: "Privacy Policy", href: "/" },
-    { label: "Help", href: "/" },
-    { label: "Tours", href: "/" },
-  ];
-
-  const contactItems: LinkItem[] = [
-    { label: "Passeio dos Descobrimentos", href: "/" },
-    { label: "8600-315 Lagos", href: "/" },
-    { label: "+999 9999 999 999", href: "/" },
-    { label: "boatripseu@gmail.com", href: "/" },
-  ];
-
   return (
     <section className={styles.contact}>
       <footer className={styles.footer}>
         <div className={styles.main}>
-          {["Quick links", "Support", "Contact info", "Connect", "Certificates"].map((title, index) => (
-            <div key={index} className={styles.list}>
-              <h4>{title}</h4>
-              <ul>
-                {title === "Connect"
-                  ? (
-                    <div className={styles.social}>
-                      <Link href={"https://www.facebook.com/boatrips.eu"} className={styles.socialLink}>
-                        <FontAwesomeIcon icon={faFacebook} />
-                      </Link>
-                      <Link href={"https://www.instagram.com/boatrips/?hl=ru"} className={styles.socialLink}>
-                        <FontAwesomeIcon icon={faInstagram} />
-                      </Link>
-                      <Link href={"/"} className={styles.socialLink}>
-                        <FontAwesomeIcon icon={faWhatsapp} />
-                      </Link>
-                    </div>
-                  )
-                  : title === "Contact info"
-                  ? (
-                    contactItems.map((item, itemIndex) => (
+          {
+            footerBlocks.map(item => Object.entries(item).map(([title, linkItems], index) =>
+                <ListWrapper title={title} key={index}>
+                  <ul>
+                     {linkItems.map((item: LinkItem, itemIndex: number) => (
                       <FooterLink key={itemIndex} label={item.label} href={item.href} />
-                    ))
-                  )
-                  : title === "Quick links" || title === "Support" ? (
-                    linkItems.map((item, itemIndex) => (
-                      <FooterLink key={itemIndex} label={item.label} href={item.href} />
-                    ))
-                  ) : (
-                    <div>
-                      <Image src="/certificates/certificate1.png" alt="certificate1" width={100} height={100} className={styles.listItem}/>
-                      <Image src="/certificates/certificate2.png" alt="certificate2" width={100} height={100} className={styles.listItem}/>
-                    </div>
-                    )}
-              </ul>
+                    ))}
+                  </ul>
+                </ListWrapper>
+            ))
+          }
+          <ListWrapper title="Connect">
+            <div className={styles.social}>
+              <Link href={"https://www.facebook.com/boatrips.eu"} className={styles.socialLink}>
+                <FontAwesomeIcon icon={faFacebook} />
+              </Link>
+              <Link href={"https://www.instagram.com/boatrips/?hl=ru"} className={styles.socialLink}>
+                <FontAwesomeIcon icon={faInstagram} />
+              </Link>
+              <Link href={"/"} className={styles.socialLink}>
+                <FontAwesomeIcon icon={faWhatsapp} />
+              </Link>
             </div>
-          ))}
+          </ListWrapper>
+          <ListWrapper title="Certificates">
+            <div className={styles.certificates}>
+              <Image src="/certificates/certificate1.png" alt="certificate1" width={120} height={120} />
+              <Image src="/certificates/certificate2.png" alt="certificate2" width={120} height={120} />
+            </div>
+          </ListWrapper>
         </div>
       </footer>
       <div className={styles.endText}>
