+++ src/components/footer/ListWrapper.tsx
@@ -1,0 +1,8 @@
+import React, { ReactNode } from 'react';
+
+export const ListWrapper = ({title, children}: {title: string, children: ReactNode}) => (
+  <div>
+    <h4>{title}</h4>
+    {children}
+  </div>
+)
