+++ src/components/footer/footer.module.scss
@@ -9,59 +9,51 @@
         display: flex;
         flex-wrap: wrap;
         justify-content: space-around;
-  
-        .list {
-          h4 {
-            font-size: 21px;
-            color: #fff;
-            margin-bottom: 30px;
-            position: relative;
-  
-            &::before {
-              content: "";
-              position: absolute;
-              height: 2px;
-              width: 60px;
-              left: 0;
-              bottom: -10px;
-              background: #fff;
-            }
+        gap: 20px;
+
+        h4 {
+          font-size: 21px;
+          color: #fff;
+          margin-bottom: 30px;
+          position: relative;
+
+          &::before {
+            content: "";
+            position: absolute;
+            height: 2px;
+            width: 60px;
+            left: 0;
+            bottom: -10px;
+            background: #fff;
           }
-  
-          ul li:not(:last-child) {
-            margin-bottom: 16px;
+        }
+
+        ul li:not(:last-child) {
+          margin-bottom: 16px;
+        }
+
+        .socialLink {
+          height: 40px;
+          width: 40px;
+          color: #166daf;
+          background-color: #fff;
+          display: inline-flex;
+          align-items: center;
+          justify-content: center;
+          font-size: 21px;
+          border-radius: 15px;
+          transition: 0.3s;
+          margin-right: 10px;
+
+          &:hover {
+            transform: scale(1.1);
           }
-  
-          .listItem {
-            color: #ffffffbf;
-            font-size: 16px;
-            display: inline-block;
-            transition: 0.3s;
-            width: inherit;
-  
-            &:hover {
-              transform: translateX(14px);
-              color: #fff;
-            }
-          }
-  
-          .socialLink {
-            height: 40px;
-            width: 40px;
-            color: #166daf;
-            background-color: #fff;
-            display: inline-flex;
-            align-items: center;
-            justify-content: center;
-            font-size: 21px;
-            border-radius: 15px;
-            transition: 0.3s;
-            margin-right: 10px;
-  
-            &:hover {
-              transform: scale(1.1);
-            }
-          }
+        }
+
+        .certificates {
+          display: flex;
+          justify-content: center;
+          gap: 20px;
         }
       }
     }
