+++ src/components/footer/footerLink/footerLink.module.scss
@@ -1,0 +1,12 @@
+.listItem {
+  color: #ffffffbf;
+  font-size: 16px;
+  display: inline-block;
+  transition: 0.3s;
+  width: inherit;
+
+  &:hover {
+    transform: translateX(14px);
+    color: #fff;
+  }
+}
